feat: Make raw template render global so resource overview can use it too
